Adapt to flattened multipart protocol
In file transfers, directories were formerly represented by
‘multipart/mixed’ content with nested boundaries.  The new flattened
multipart file uploads introduced in IPFS 0.4.0 now use the content-
type ‘application/x-directory’.

Furthermore, I gave up on implementing ReadFromWriterIO as a sub-class
of IO, since it got broken with Ruby 2.4.0 (copy_stream).  It is
replaced by ‘StreamProducer’ which simply writes to a temp file
and then returns a reader on it.